Move test_utils to tests folder
This remove the file from coverage.
